+++ lib/api/entities.rb
@@ -50,7 +50,7 @@ module API
       expose :project_id, :push_events
       expose :issues_events, :merge_requests_events, :tag_push_events
       expose :note_events, :build_events, :pipeline_events, :wiki_page_events
-      expose :enable_ssl_verification
+      expose :enable_ssl_verification, :token
     end
 
     class BasicProjectDetails < Grape::Entity